Implement network.fqdns module function
